requirements.txt
torch>=2.0.0
torchvision>=0.15.0
Pillow>=10.0.0
numpy>=1.24.0
scikit-learn>=1.0.0
pandas>=1.3.0
matplotlib>=3.4.0
tqdm>=4.65.0
python-dotenv>=0.19.0
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
flask>=2.0.0
flask-cors>=4.0.0
werkzeug>=2.0.0
firebase-admin>=6.0.0 